Added second plot (of all individual obs. and sim. annual coffee yields) to BC.AverageYields.R

+++ BC_CAF2021_Masatepe_01-14.R
@@ -3,7 +3,7 @@
 
 ## 1. INITIALISE MCMC ##
    # nChain <- as.integer(36000)
-   nChain <- as.integer(500)
+   nChain <- as.integer(1e5)
    source('BC/BC_CAF2021_MCMC_init_Masatepe_01-14.R')
 
 ## 2. RUNNING THE MCMC ##
